requirements.txt
openai==1.95.1
pandas==2.3.1
python-dotenv==1.1.1
requests==2.32.4
google-genai==1.25.0
